separacao ,de um código recorrente, para um arquivo separado
